+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/api2/model/dto/MessageDtos.kt
@@ -128,4 +128,5 @@ internal data class DownstreamDraftMessageDto(
 @JsonClass(generateAdapter = true)
 internal data class DownstreamPendingMessageDto(
     val message: DownstreamMessageDto,
+    val metadata: Map<String, String>?,
 )