+++ stream-chat-android-client/src/test/java/io/getstream/chat/android/client/parser2/testdata/ChannelDtoTestData.kt
@@ -63,7 +63,8 @@ internal object ChannelDtoTestData {
            }
           ],
           "user_message_reminders": false,
-          "shared_locations": true
+          "shared_locations": true,
+          "mark_messages_pending": false
         }
         """.withoutWhitespace()
 
@@ -99,6 +100,7 @@ internal object ChannelDtoTestData {
         ),
         user_message_reminders = false,
         shared_locations = true,
+        mark_messages_pending = false,
     )
 
     @Language("JSON")