+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/repository/domain/channelconfig/internal/ChannelConfigEntity.kt
@@ -48,6 +48,7 @@ internal data class ChannelConfigInnerEntity(
     val automodBehavior: String,
     val blocklistBehavior: String,
     val messageRemindersEnabled: Boolean,
+    val markMessagesPending: Boolean,
 )
 
 @Entity(